requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.1.26
langchain-openai==0.0.7
langsmith==0.1.8
linkpreview==0.8.3
neonize==0.0.7
numpy==1.26.4
openai==1.12.0
orjson==3.9.15
packaging==23.2
phonenumbers==8.13.29
pillow==10.2.0
protobuf==4.25.2
psutil==5.9.8
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
segno==1.6.0
sniffio==1.3.0
soupsieve==2.5
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.6.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
watchdog==3.0.0
youtube-search==2.1.2
